updates index and state for states to have get, put, post, and delete, methods
