+++ api/v1/views/index.py
@@ -31,42 +31,3 @@ def display_stats():
         "users": users
     }
     return jsonify(stats)
-
-@app_views.route('/states/', methods = ['POST', 'GET', 'PUT', 'DELETE'])
-@app_views.route('/states/<id>', methods = ['POST', 'GET', 'PUT', 'DELETE'])
-def get_states(id=None):
-    '''
-        This will ...
-    '''
-    if request.method == 'GET':
-        result = states.get_method(id)
-        if result == None:
-            abort(404)
-        return jsonify(result)
-
-    elif request.method == 'POST':
-        if not request.json:
-            abort(400, "Not a JSON")
-        if "name" not in request.json:
-            abort(400, "Missing name")
-        body = request.get_json()
-        result = states.post_method(body)
-        return jsonify(result), 201
-
-    elif request.method == 'PUT':
-        if not request.json:
-            abort(400, "Not a JSON")
-        if "name" not in request.json:
-            abort(400, "Missing name")
-        body = request.get_json()
-        result = states.put_method(id, body)
-        if result == None:
-            abort(404)
-        return jsonify(result), 200
-    elif request.method == 'DELETE':
-        result = states.delete_method(id)
-        if result == None:
-            abort(404)
-        return jsonify(result), 200
-    else:
-        pass
